+++ clang/test/CodeGen/pragma-cx-limited-range.c
@@ -2,13 +2,13 @@
 // RUN: -o - | FileCheck %s --check-prefix=FULL
 
 // RUN: %clang_cc1 %s -O0 -emit-llvm -triple x86_64-unknown-unknown \
-// RUN: -fcx-limited-range -o - | FileCheck --check-prefix=LMTD %s
+// RUN: -complex-range=cx_limited -o - | FileCheck --check-prefix=LMTD %s
 
 // RUN: %clang_cc1 %s -O0 -emit-llvm -triple x86_64-unknown-unknown \
 // RUN: -fno-cx-limited-range -o - | FileCheck %s --check-prefix=FULL
 
 // RUN: %clang_cc1 %s -O0 -emit-llvm -triple x86_64-unknown-unknown \
-// RUN: -fcx-fortran-rules -o - | FileCheck --check-prefix=FRTRN %s
+// RUN: -complex-range=cx_fortran -o - | FileCheck --check-prefix=FRTRN %s
 
 // RUN: %clang_cc1 %s -O0 -emit-llvm -triple x86_64-unknown-unknown \
 // RUN: -fno-cx-fortran-rules -o - | FileCheck --check-prefix=FULL %s
@@ -45,12 +45,7 @@ _Complex float pragma_off_mul(_Complex float a, _Complex float b) {
   // LABEL: define {{.*}} @pragma_off_mul(
   // FULL: call {{.*}} @__mulsc3
 
-  // LMTD: fmul float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fsub float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fadd float
+  // LMTD: call {{.*}} @__mulsc3
 
   // FRTRN: call {{.*}} @__mulsc3
 
@@ -104,47 +99,9 @@ _Complex float pragma_off_div(_Complex float a, _Complex float b) {
   // LABEL: define {{.*}} @pragma_off_div(
   // FULL: call {{.*}} @__divsc3
 
-  // LMTD: fmul float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fadd float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fadd float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fmul float
-  // LMTD-NEXT: fsub float
-  // LMTD-NEXT: fdiv float
-  // LMTD-NEXT: fdiv float
+  // LMTD: call {{.*}} @__divsc3
 
-  // FRTRN: call float @llvm.fabs.f32(float {{.*}})
-  // FRTRN-NEXT: call float @llvm.fabs.f32(float {{.*}})
-  // FRTRN-NEXT: fcmp ugt float
-  // FRTRN-NEXT: br i1 {{.*}}, label
-  // FRTRN:      true_bb_name:
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fadd float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fadd float
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fsub float
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: br label
-  // FRTRN:      false_bb_name:
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fadd float
-  // FRTRN-NEXT: fadd float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: fsub float
-  // FRTRN-NEXT: fmul float
-  // FRTRN-NEXT: fdiv float
-  // FRTRN-NEXT: br label
-  // FRTRN:      cont_bb:
-  // FRTRN-NEXT: phi float
-  // FRTRN-NEXT: phi float
+  // FRTRN: call {{.*}} @__divsc3
 
   return a / b;
 }